feat(ux): IAM gold standard — internal menus, absorb 6 scripts — closes #74
